Mostly finished. Segfaults when calling the C libraries, probably due to file pointers or array pointers. Sort these out by systematically removing degrees of freedom in a throwaway C routine until the offenders are identified.
For videos, add a video = params key which tells us what parameters we want to follow. Make a subdirectory for each, and have a different file for each time step. These then need their own colormap schemes, since values can be positive or negative. For angles, use the phase mapper from lattice_monte_carlo; technicolor videos should cause people to trip balls. For time, this can be really cool: plot the image as it progresses, with time as the color, and halt evolution in a pixel when the rule is satisfied; we can then watch the normal images form. This will require obscene amounts of hard disk space, but we could even zip the images on the fly to save room.